refactor(ui): isolate chat stream reconciliation
