+++ ui/src/ui/chat/stream-text.ts
@@ -1,0 +1,6 @@
+export function trimAccumulatedStreamPrefix(text: string, previousText: string | null): string {
+  if (!previousText || !text.startsWith(previousText)) {
+    return text;
+  }
+  return text.slice(previousText.length).trimStart();
+}
