+++ ui/src/ui/controllers/chat.ts
@@ -1,4 +1,3 @@
-import { resetToolStream } from "../app-tool-stream.ts";
 import { getChatAttachmentDataUrl } from "../chat/attachment-payload-store.ts";
 import {
   isAssistantHeartbeatAckForDisplay,
@@ -6,7 +5,18 @@ import {
 } from "../chat/heartbeat-display.ts";
 import { extractText } from "../chat/message-extract.ts";
 import { reconcileChatRunLifecycle } from "../chat/run-lifecycle.ts";
-import { extractToolMessageRefs } from "../chat/tool-message-refs.ts";
+import {
+  appendTerminalAssistantMessage,
+  clearToolStreamSegments,
+  currentLiveToolCallIds,
+  hasVisibleStreamParts,
+  historyReplacedVisibleStream,
+  materializeVisibleStreamState,
+  maybeResetToolStream,
+  persistedCurrentToolStreamIds,
+  prunePersistedToolStreamMessages,
+  visibleCurrentAssistantStreamTail,
+} from "../chat/stream-reconciliation.ts";
 import { buildUserChatMessageContentBlocks } from "../chat/user-message-content.ts";
 import { formatConnectError } from "../connect-error.ts";
 import {
@@ -96,36 +106,6 @@ function isSyntheticTranscriptRepairToolResult(message: unknown): boolean {
   return typeof text === "string" && text.trim() === SYNTHETIC_TRANSCRIPT_REPAIR_RESULT;
 }
 
-function currentLiveToolCallIds(state: ChatState): string[] {
-  const toolHost = state as ChatState & { toolStreamOrder?: unknown };
-  return Array.isArray(toolHost.toolStreamOrder)
-    ? toolHost.toolStreamOrder.filter(
-        (value): value is string => typeof value === "string" && value.trim().length > 0,
-      )
-    : [];
-}
-
-function persistedCurrentToolStreamIds(messages: unknown[], state: ChatState): Set<string> {
-  const liveToolIds = currentLiveToolCallIds(state);
-  const matchedToolIds = new Set<string>();
-  if (liveToolIds.length === 0) {
-    return matchedToolIds;
-  }
-  const liveToolIdSet = new Set(liveToolIds);
-  const persistedToolIds = new Set<string>();
-  for (const message of messages.slice(lastUserMessageIndex(messages) + 1)) {
-    for (const ref of extractToolMessageRefs(message)) {
-      persistedToolIds.add(ref.id);
-    }
-  }
-  for (const id of persistedToolIds) {
-    if (liveToolIdSet.has(id)) {
-      matchedToolIds.add(id);
-    }
-  }
-  return matchedToolIds;
-}
-
 function isTextOnlyContent(content: unknown): boolean {
   if (typeof content === "string") {
     return true;
@@ -179,6 +159,10 @@ function isHeartbeatAckStream(text: string): boolean {
   return stripHeartbeatTokenForDisplay(text).shouldSkip;
 }
 
+function isHiddenAssistantStreamText(text: string): boolean {
+  return isSilentReplyStream(text) || isHeartbeatAckStream(text);
+}
+
 function shouldHideAssistantChatMessage(message: unknown): boolean {
   return isAssistantSilentReply(message) || isAssistantHeartbeatAckForDisplay(message);
 }
@@ -189,6 +173,22 @@ function shouldHideHistoryMessage(message: unknown): boolean {
     isSyntheticTranscriptRepairToolResult(message) ||
     isEmptyUserTextOnlyMessage(message)
   );
+}
+
+function materializeVisibleAssistantStreamMessages(
+  messages: unknown[],
+  state: ChatState,
+  opts: {
+    includeCurrent?: boolean;
+    requirePersistedTool?: boolean;
+    replacementMessages?: unknown[];
+  } = {},
+): unknown[] {
+  return materializeVisibleStreamState(messages, state, {
+    ...opts,
+    isHiddenAssistantMessage: shouldHideAssistantChatMessage,
+    isHiddenStreamText: isHiddenAssistantStreamText,
+  });
 }
 
 function hasTranscriptMeta(message: unknown): boolean {
@@ -226,286 +226,6 @@ function messageDisplaySignature(message: unknown): string | null {
   } catch {
     return null;
   }
-}
-
-function visibleAssistantStreamText(stream: string | null): string | null {
-  if (!stream?.trim() || isSilentReplyStream(stream) || isHeartbeatAckStream(stream)) {
-    return null;
-  }
-  return stream;
-}
-
-function buildAssistantStreamMessage(
-  stream: string,
-  replacementText = stream,
-  timestamp = Date.now(),
-): Record<string, unknown> {
-  return {
-    role: "assistant",
-    content: [{ type: "text", text: stream }],
-    timestamp,
-    openclawStreamFallback: {
-      replacementText,
-    },
-  };
-}
-
-function streamFallbackReplacementText(message: unknown): string | null {
-  if (!message || typeof message !== "object") {
-    return null;
-  }
-  const fallback = (message as { openclawStreamFallback?: unknown }).openclawStreamFallback;
-  if (!fallback || typeof fallback !== "object") {
-    return null;
-  }
-  const replacementText = (fallback as { replacementText?: unknown }).replacementText;
-  if (typeof replacementText === "string" && replacementText.trim()) {
-    return replacementText.trim();
-  }
-  return extractText(message)?.trim() ?? null;
-}
-
-function terminalMessageReplacesStreamFallback(message: unknown, fallback: unknown): boolean {
-  const fallbackText = streamFallbackReplacementText(fallback);
-  if (!fallbackText) {
-    return false;
-  }
-  const terminalText = extractText(message)?.trim();
-  return Boolean(
-    terminalText && (terminalText === fallbackText || terminalText.startsWith(fallbackText)),
-  );
-}
-
-function appendTerminalAssistantMessage(messages: unknown[], message: unknown): unknown[] {
-  const retainedMessages = messages.filter((existing, index) => {
-    if (index <= lastUserMessageIndex(messages)) {
-      return true;
-    }
-    return !terminalMessageReplacesStreamFallback(message, existing);
-  });
-  return [...retainedMessages, message];
-}
-
-function lastUserMessageIndex(messages: unknown[]): number {
-  for (let index = messages.length - 1; index >= 0; index--) {
-    const message = messages[index];
-    if (!message || typeof message !== "object") {
-      continue;
-    }
-    const role = normalizeLowercaseStringOrEmpty((message as { role?: unknown }).role);
-    if (role === "user") {
-      return index;
-    }
-  }
-  return -1;
-}
-
-function hasAssistantStreamReplacement(messages: unknown[], stream: string): boolean {
-  const expected = stream.trim();
-  if (!expected) {
-    return false;
-  }
-  const startIndex = lastUserMessageIndex(messages) + 1;
-  return messages.slice(startIndex).some((message) => {
-    if (!message || typeof message !== "object") {
-      return false;
-    }
-    const role = normalizeLowercaseStringOrEmpty((message as { role?: unknown }).role);
-    if (role && role !== "assistant") {
-      return false;
-    }
-    if (role === "assistant" && shouldHideAssistantChatMessage(message)) {
-      return false;
-    }
-    const text = extractText(message)?.trim();
-    return Boolean(text && (text === expected || text.startsWith(expected)));
-  });
-}
-
-function trimAccumulatedVisibleStreamText(text: string, previousText: string | null): string {
-  if (!previousText || !text.startsWith(previousText)) {
-    return text;
-  }
-  return text.slice(previousText.length).trimStart();
-}
-
-type VisibleAssistantStreamPart = {
-  text: string;
-  replacementText: string;
-  source: "segment" | "current";
-  timestamp: number;
-  toolCallId?: string;
-};
-
-function visibleAssistantStreamParts(
-  state: ChatState,
-  opts?: { includeCurrent?: boolean },
-): VisibleAssistantStreamPart[] {
-  const streamHost = state as ChatState & {
-    chatStreamSegments?: Array<{ text?: unknown; ts?: unknown; toolCallId?: unknown }>;
-  };
-  const liveToolIds = currentLiveToolCallIds(state);
-  const parts: VisibleAssistantStreamPart[] = [];
-  let previousText: string | null = null;
-  const segments = Array.isArray(streamHost.chatStreamSegments)
-    ? streamHost.chatStreamSegments
-    : [];
-  for (let segmentIndex = 0; segmentIndex < segments.length; segmentIndex++) {
-    const segment = segments[segmentIndex];
-    if (!segment) {
-      continue;
-    }
-    if (typeof segment.text !== "string") {
-      continue;
-    }
-    const visible = visibleAssistantStreamText(
-      trimAccumulatedVisibleStreamText(segment.text, previousText),
-    );
-    if (visible) {
-      parts.push({
-        text: visible,
-        replacementText: segment.text,
-        source: "segment",
-        timestamp:
-          typeof segment.ts === "number" && Number.isFinite(segment.ts) ? segment.ts : Date.now(),
-        toolCallId:
-          typeof segment.toolCallId === "string" && segment.toolCallId.trim()
-            ? segment.toolCallId.trim()
-            : liveToolIds[segmentIndex],
-      });
-    }
-    if (segment.text.trim()) {
-      previousText = segment.text;
-    }
-  }
-  if (opts?.includeCurrent !== false && typeof state.chatStream === "string") {
-    const visible = visibleAssistantStreamText(
-      trimAccumulatedVisibleStreamText(state.chatStream, previousText),
-    );
-    if (visible) {
-      parts.push({
-        text: visible,
-        replacementText: state.chatStream,
-        source: "current",
-        timestamp: state.chatStreamStartedAt ?? Date.now(),
-      });
-    }
-  }
-  return parts;
-}
-
-function visibleCurrentAssistantStreamTail(state: ChatState): string | null {
-  if (typeof state.chatStream !== "string") {
-    return null;
-  }
-  const streamHost = state as ChatState & {
-    chatStreamSegments?: Array<{ text?: unknown }>;
-  };
-  const segments = Array.isArray(streamHost.chatStreamSegments)
-    ? streamHost.chatStreamSegments
-    : [];
-  let previousText: string | null = null;
-  for (const segment of segments) {
-    if (typeof segment.text === "string" && segment.text.trim()) {
-      previousText = segment.text;
-    }
-  }
-  return visibleAssistantStreamText(
-    trimAccumulatedVisibleStreamText(state.chatStream, previousText),
-  );
-}
-
-function hasAssistantStreamPartReplacement(
-  messages: unknown[],
-  part: VisibleAssistantStreamPart,
-): boolean {
-  return (
-    hasAssistantStreamReplacement(messages, part.replacementText) ||
-    hasAssistantStreamReplacement(messages, part.text)
-  );
-}
-
-function currentToolStreamMessageIndex(
-  messages: unknown[],
-  state: ChatState,
-  toolCallId?: string,
-): number {
-  const liveToolIds = toolCallId ? new Set([toolCallId]) : new Set(currentLiveToolCallIds(state));
-  if (liveToolIds.size === 0) {
-    return -1;
-  }
-  const startIndex = lastUserMessageIndex(messages) + 1;
-  for (let index = startIndex; index < messages.length; index++) {
-    if (extractToolMessageRefs(messages[index]).some((ref) => liveToolIds.has(ref.id))) {
-      return index;
-    }
-  }
-  return -1;
-}
-
-function insertMessageAtIndex(messages: unknown[], message: unknown, index: number): unknown[] {
-  return [...messages.slice(0, index), message, ...messages.slice(index)];
-}
-
-function timestampForInsertedVisibleStream(
-  messages: unknown[],
-  index: number,
-  desiredTimestamp: number,
-): number {
-  const previousTimestamp = messages
-    .slice(0, index)
-    .toReversed()
-    .map(messageTimestampMs)
-    .find((timestamp): timestamp is number => timestamp != null);
-  const nextTimestamp = messages
-    .slice(index)
-    .map(messageTimestampMs)
-    .find((timestamp): timestamp is number => timestamp != null);
-  if (previousTimestamp != null && desiredTimestamp <= previousTimestamp) {
-    const afterPrevious = previousTimestamp + 1;
-    return nextTimestamp != null && afterPrevious >= nextTimestamp
-      ? previousTimestamp + (nextTimestamp - previousTimestamp) / 2
-      : afterPrevious;
-  }
-  if (nextTimestamp != null && desiredTimestamp >= nextTimestamp) {
-    const beforeNext = nextTimestamp - 1;
-    return previousTimestamp != null && beforeNext <= previousTimestamp
-      ? previousTimestamp + (nextTimestamp - previousTimestamp) / 2
-      : beforeNext;
-  }
-  return desiredTimestamp;
-}
-
-function appendVisibleStreamStateMessages(
-  messages: unknown[],
-  state: ChatState,
-  replacementMessages: unknown[] = [],
-  opts?: { includeCurrent?: boolean; requirePersistedTool?: boolean },
-): unknown[] {
-  let nextMessages = messages;
-  for (const part of visibleAssistantStreamParts(state, opts)) {
-    if (hasAssistantStreamPartReplacement([...nextMessages, ...replacementMessages], part)) {
-      continue;
-    }
-    const toolIndex =
-      part.source === "segment"
-        ? currentToolStreamMessageIndex(nextMessages, state, part.toolCallId)
-        : -1;
-    if (opts?.requirePersistedTool && toolIndex < 0) {
-      continue;
-    }
-    const insertIndex = toolIndex >= 0 ? toolIndex : nextMessages.length;
-    const streamMessage = buildAssistantStreamMessage(
-      part.text,
-      part.replacementText,
-      timestampForInsertedVisibleStream(nextMessages, insertIndex, part.timestamp),
-    );
-    nextMessages =
-      toolIndex >= 0
-        ? insertMessageAtIndex(nextMessages, streamMessage, toolIndex)
-        : [...nextMessages, streamMessage];
-  }
-  return nextMessages;
 }
 
 function messageTimestampMs(message: unknown): number | null {
@@ -795,82 +515,6 @@ function chatEventSessionMatches(state: ChatState, payload: ChatEventPayload): b
   );
 }
 
-function maybeResetToolStream(state: ChatState, opts?: { preserveStreamSegments?: boolean }) {
-  const toolHost = state as ChatState & Partial<Parameters<typeof resetToolStream>[0]>;
-  if (
-    toolHost.toolStreamById instanceof Map &&
-    Array.isArray(toolHost.toolStreamOrder) &&
-    Array.isArray(toolHost.chatToolMessages) &&
-    Array.isArray(toolHost.chatStreamSegments)
-  ) {
-    const preservedStreamSegments = opts?.preserveStreamSegments
-      ? [...toolHost.chatStreamSegments]
-      : null;
-    resetToolStream(toolHost as Parameters<typeof resetToolStream>[0]);
-    if (preservedStreamSegments) {
-      toolHost.chatStreamSegments = preservedStreamSegments;
-    }
-  }
-}
-
-function clearToolStreamSegments(state: ChatState) {
-  const toolHost = state as ChatState & { chatStreamSegments?: unknown };
-  if (Array.isArray(toolHost.chatStreamSegments)) {
-    toolHost.chatStreamSegments = [];
-  }
-}
-
-function prunePersistedToolStreamMessages(state: ChatState, persistedToolIds: Set<string>) {
-  if (persistedToolIds.size === 0) {
-    return;
-  }
-  const toolHost = state as ChatState & {
-    chatStreamSegments?: Array<{ text?: unknown; ts?: unknown; toolCallId?: unknown }>;
-    chatToolMessages?: unknown[];
-    toolStreamById?: Map<string, unknown>;
-    toolStreamOrder?: unknown[];
-  };
-  const liveToolIds = currentLiveToolCallIds(state);
-  if (toolHost.toolStreamById instanceof Map) {
-    for (const id of persistedToolIds) {
-      toolHost.toolStreamById.delete(id);
-    }
-  }
-  if (Array.isArray(toolHost.toolStreamOrder)) {
-    toolHost.toolStreamOrder = toolHost.toolStreamOrder.filter(
-      (id): id is string => typeof id === "string" && !persistedToolIds.has(id),
-    );
-  }
-  if (Array.isArray(toolHost.chatToolMessages)) {
-    toolHost.chatToolMessages = toolHost.chatToolMessages.filter((message) => {
-      const refs = extractToolMessageRefs(message);
-      return refs.every((ref) => !persistedToolIds.has(ref.id));
-    });
-  }
-  if (!Array.isArray(toolHost.chatStreamSegments)) {
-    return;
-  }
-  let lastPrunedAccumulatedText: string | null = null;
-  toolHost.chatStreamSegments = toolHost.chatStreamSegments.flatMap((segment, index) => {
-    const explicitToolCallId =
-      typeof segment.toolCallId === "string" && segment.toolCallId.trim()
-        ? segment.toolCallId.trim()
-        : null;
-    const toolCallId = explicitToolCallId ?? liveToolIds[index] ?? null;
-    const text = typeof segment.text === "string" ? segment.text : "";
-    if (toolCallId && persistedToolIds.has(toolCallId)) {
-      if (text.trim()) {
-        lastPrunedAccumulatedText = text;
-      }
-      return [];
-    }
-    const nextText = lastPrunedAccumulatedText
-      ? trimAccumulatedVisibleStreamText(text, lastPrunedAccumulatedText)
-      : text;
-    return [{ ...segment, text: nextText }];
-  });
-}
-
 function resolveDeltaChatStreamText(
   currentStream: string | null,
   payload: ChatEventPayload,
@@ -1084,19 +728,23 @@ async function loadChatHistoryUncached(
     state.chatThinkingLevel = res.sessionInfo?.thinkingLevel ?? res.thinkingLevel ?? null;
     const resetStream = !state.chatRunId || state.chatRunId === previousRunId;
     if (resetStream) {
-      const visibleStreamParts = visibleAssistantStreamParts(state);
-      const historyReplacedStream =
-        visibleStreamParts.length > 0 &&
-        visibleStreamParts.every((part) =>
-          hasAssistantStreamPartReplacement(state.chatMessages, part),
-        );
+      const streamReconciliation = {
+        isHiddenAssistantMessage: shouldHideAssistantChatMessage,
+        isHiddenStreamText: isHiddenAssistantStreamText,
+      };
+      const hasVisibleStream = hasVisibleStreamParts(state, streamReconciliation);
+      const historyReplacedStream = historyReplacedVisibleStream(
+        state.chatMessages,
+        state,
+        streamReconciliation,
+      );
       const liveToolIds = currentLiveToolCallIds(state);
       const persistedToolStreamIds = persistedCurrentToolStreamIds(state.chatMessages, state);
       const historyReplacedToolStream =
         liveToolIds.length > 0 && liveToolIds.every((id) => persistedToolStreamIds.has(id));
       const historyReplacedSomeToolStream = persistedToolStreamIds.size > 0;
       const liveToolStreamReplaced = liveToolIds.length === 0 || historyReplacedToolStream;
-      if (visibleStreamParts.length === 0 || historyReplacedStream) {
+      if (!hasVisibleStream || historyReplacedStream) {
         if (liveToolStreamReplaced) {
           // Clear all streaming state — history includes tool results and text
           // inline, so keeping streaming artifacts would cause duplicates.
@@ -1115,22 +763,28 @@ async function loadChatHistoryUncached(
           visibleMessageCount: visibleMessages.length,
         });
       } else if (!state.chatRunId) {
-        state.chatMessages = appendVisibleStreamStateMessages(state.chatMessages, state);
+        state.chatMessages = materializeVisibleAssistantStreamMessages(state.chatMessages, state);
         maybeResetToolStream(state);
         state.chatStream = null;
         state.chatStreamStartedAt = null;
       } else if (historyReplacedToolStream) {
-        state.chatMessages = appendVisibleStreamStateMessages(state.chatMessages, state, [], {
+        state.chatMessages = materializeVisibleAssistantStreamMessages(state.chatMessages, state, {
           includeCurrent: false,
         });
-        state.chatStream = visibleCurrentAssistantStreamTail(state);
+        state.chatStream = visibleCurrentAssistantStreamTail(
+          state,
+          streamReconciliation.isHiddenStreamText,
+        );
         if (state.chatStream === null) {
           state.chatStreamStartedAt = null;
         }
         maybeResetToolStream(state);
       } else if (historyReplacedSomeToolStream) {
-        const visibleCurrentTail = visibleCurrentAssistantStreamTail(state);
-        state.chatMessages = appendVisibleStreamStateMessages(state.chatMessages, state, [], {
+        const visibleCurrentTail = visibleCurrentAssistantStreamTail(
+          state,
+          streamReconciliation.isHiddenStreamText,
+        );
+        state.chatMessages = materializeVisibleAssistantStreamMessages(state.chatMessages, state, {
           includeCurrent: false,
           requirePersistedTool: true,
         });
@@ -1560,21 +1214,19 @@ export function handleChatEvent(state: ChatState, payload?: ChatEventPayload) {
     if (finalMessage && !shouldHideAssistantChatMessage(finalMessage)) {
       state.chatMessages = appendTerminalAssistantMessage(state.chatMessages, finalMessage);
     } else {
-      state.chatMessages = appendVisibleStreamStateMessages(state.chatMessages, state);
+      state.chatMessages = materializeVisibleAssistantStreamMessages(state.chatMessages, state);
     }
     reconcileTerminalRun("done", "done");
   } else if (payload.state === "aborted") {
     const normalizedMessage = normalizeAbortedAssistantMessage(payload.message);
     if (normalizedMessage && !shouldHideAssistantChatMessage(normalizedMessage)) {
-      state.chatMessages = appendVisibleStreamStateMessages(
-        state.chatMessages,
-        state,
-        [normalizedMessage],
-        { includeCurrent: false },
-      );
+      state.chatMessages = materializeVisibleAssistantStreamMessages(state.chatMessages, state, {
+        replacementMessages: [normalizedMessage],
+        includeCurrent: false,
+      });
       state.chatMessages = appendTerminalAssistantMessage(state.chatMessages, normalizedMessage);
     } else {
-      state.chatMessages = appendVisibleStreamStateMessages(state.chatMessages, state);
+      state.chatMessages = materializeVisibleAssistantStreamMessages(state.chatMessages, state);
     }
     reconcileTerminalRun("interrupted", "killed");
   } else if (payload.state === "error") {
@@ -1584,9 +1236,9 @@ export function handleChatEvent(state: ChatState, payload?: ChatEventPayload) {
     const visiblePayloadMessage =
       payloadMessage && !shouldHideAssistantChatMessage(payloadMessage) ? payloadMessage : null;
     if (visiblePayloadMessage) {
-      state.chatMessages = appendVisibleStreamStateMessages(state.chatMessages, state, [
-        visiblePayloadMessage,
-      ]);
+      state.chatMessages = materializeVisibleAssistantStreamMessages(state.chatMessages, state, {
+        replacementMessages: [visiblePayloadMessage],
+      });
       state.chatMessages = appendTerminalAssistantMessage(
         state.chatMessages,
         visiblePayloadMessage,
@@ -1594,7 +1246,7 @@ export function handleChatEvent(state: ChatState, payload?: ChatEventPayload) {
     } else {
       const errorMessage = hadActiveRunBeforeEvent ? buildErrorAssistantMessage(payload) : null;
       if (hadActiveRunBeforeEvent) {
-        state.chatMessages = appendVisibleStreamStateMessages(state.chatMessages, state);
+        state.chatMessages = materializeVisibleAssistantStreamMessages(state.chatMessages, state);
       }
       if (errorMessage) {
         state.chatMessages = appendTerminalAssistantMessage(state.chatMessages, errorMessage);
